Provide atomic interface for cache helper
This introduces an alternative method for the cache helper that allows
reducing the public API and removes the need for coordinating method
calls.

It also reduces the amount of method calls needed to use the object,
which is usually irrelevant for small projects but should show
improvements when scanning several files.

Signed-off-by: Luís Cobucci <lcobucci@gmail.com>